`IntervalSnipper` doesn't invoke `notify()` if OCR fails
Splits `utils.py` into `ocr.py`, `logger.py` and `notifications.py`.
