+++ messages.py
@@ -1,0 +1,5 @@
+ocr_failure_message = "Unable to read text from image, did not copy"
+
+
+def ocr_error_message(
+    error): return "An error occurred when trying to process the image: {error}"
